Animation component: extracted authoring html and controller functions to separate authoring files. Closes #1255
