+++ src/main/webapp/wise5/components/animation/animationAuthoringComponentModule.es6
@@ -2,17 +2,18 @@
 
 import AnimationService from './animationService';
 import AnimationController from './animationController';
+import AnimationAuthoringController from './animationAuthoringController';
 
-const animationComponentModule = angular.module('animationComponentModule', [
+const animationAuthoringComponentModule = angular.module('animationAuthoringComponentModule', [
     'pascalprecht.translate'
   ])
   .service(AnimationService.name, AnimationService)
   .controller(AnimationController.name, AnimationController)
+  .controller(AnimationAuthoringController.name, AnimationAuthoringController)
   .config([
     '$translatePartialLoaderProvider',
     ($translatePartialLoaderProvider) => {
       $translatePartialLoaderProvider.addPart('components/animation/i18n');
     }
   ]);
-
-export default animationComponentModule;
+export default animationAuthoringComponentModule;
