+++ src/main/webapp/wise5/directives/component/component.es6
@@ -77,10 +77,9 @@ class ComponentController {
             $scope.authoringComponentContent = authoringComponentContent;
             $scope.authoringComponentContentJSONString = angular.toJson($scope.authoringComponentContent, 4);
             $scope.nodeAuthoringController = $scope.$parent.nodeAuthoringController;
-
         }
 
-        if ($scope.mode === 'authoring' && (componentContent.type == 'HTML' || componentContent.type == 'ConceptMap')) {
+        if ($scope.mode === 'authoring' && (componentContent.type == 'HTML' || componentContent.type == 'ConceptMap' || componentContent.type == 'Animation')) {
             $scope.componentTemplatePath = this.NodeService.getComponentAuthoringTemplatePath(componentContent.type);
         } else {
             $scope.componentTemplatePath = this.NodeService.getComponentTemplatePath(componentContent.type);
